+++ packages/charting/src/components/StackedBarChart/StackedBarChart.base.tsx
@@ -29,8 +29,7 @@ export class StackedBarChartBase extends React.Component<IStackedBarChartProps, 
   public static defaultProps: Partial<IStackedBarChartProps> = {
     barHeight: 16,
     hideNumberDisplay: false,
-    hideLegend: false,
-    isMultiStackedBarChart: false
+    hideLegend: false
   };
   private _classNames: IProcessedStyleSet<IStackedBarChartStyles>;
 
@@ -53,9 +52,10 @@ export class StackedBarChartBase extends React.Component<IStackedBarChartProps, 
 
   public render(): JSX.Element {
     this._adjustProps();
-    const { data, barHeight, hideNumberDisplay, hideLegend, theme, styles, barBackgroundColor } = this.props;
+    const { data, hideNumberDisplay, hideLegend, theme, styles, barBackgroundColor, href } = this.props;
     const { palette } = theme!;
-    const bars = this._createBarsAndLegends(data!, barHeight!, palette, barBackgroundColor);
+    const barHeight = data!.chartData!.length > 2 ? this.props.barHeight : 8;
+    const bars = this._createBarsAndLegends(data!, barHeight!, palette, barBackgroundColor, href);
     const showRatio = hideNumberDisplay === false && data!.chartData!.length === 2;
     const showNumber = hideNumberDisplay === false && data!.chartData!.length === 1;
     let total = 0;
@@ -78,7 +78,8 @@ export class StackedBarChartBase extends React.Component<IStackedBarChartProps, 
           )}
           {showRatio && (
             <div>
-              <strong>{data!.chartData![0].data}</strong>/{total}
+              <span className={this._classNames.ratioNumerator}>{data!.chartData![0].data}</span>/
+              <span className={this._classNames.ratioDenominator}>{total}</span>
             </div>
           )}
           {showNumber && (
@@ -125,7 +126,8 @@ export class StackedBarChartBase extends React.Component<IStackedBarChartProps, 
     data: IChartProps,
     barHeight: number,
     palette: IPalette,
-    barBackgroundColor?: string
+    barBackgroundColor?: string,
+    href?: string
   ): [JSX.Element[], JSX.Element] {
     const defaultPalette: string[] = [palette.blueLight, palette.blue, palette.blueMid, palette.red, palette.black];
     const legendDataItems: ILegend[] = [];
@@ -173,7 +175,8 @@ export class StackedBarChartBase extends React.Component<IStackedBarChartProps, 
       }
       this._classNames = getClassNames(styles!, {
         theme: this.props.theme!,
-        shouldHighlight: shouldHighlight
+        shouldHighlight: shouldHighlight,
+        href: href
       });
       return (
         <g
@@ -186,6 +189,7 @@ export class StackedBarChartBase extends React.Component<IStackedBarChartProps, 
           onMouseMove={this._onBarHover.bind(this, point.legend!, pointData, color)}
           onMouseLeave={this._onBarLeave}
           pointerEvents="all"
+          onClick={this._redirectToUrl.bind(this, href)}
         >
           <rect key={index} x={startingPoint[index] + '%'} y={0} width={value + '%'} height={barHeight} fill={color} />
         </g>
@@ -267,5 +271,9 @@ export class StackedBarChartBase extends React.Component<IStackedBarChartProps, 
     this.setState({
       isCalloutVisible: false
     });
+  }
+
+  private _redirectToUrl(href: string | undefined): void {
+    href ? (window.location.href = href) : '';
   }
 }