+++ packages/dashboard/src/components/Card/Card.types.ts
@@ -99,6 +99,16 @@ export interface ICardFrameContent {
    * Hyperlink URL for title
    */
   href?: string;
+
+  /**
+   * Target for Hyperlink URL for title
+   */
+  target?: string;
+
+  /**
+   * callback triggered upon clicking on the card title. Card title is clickable only when href is passed to it.
+   */
+  cardTitleCallback?: VoidFunction;
 }
 
 export interface ICardProps {