+++ packages/experiments/src/components/Stack/VerticalStack/VerticalStack.tsx
@@ -1,32 +1,33 @@
 import * as React from 'react';
-import { createStatelessComponent, IViewComponentProps } from '../../../Foundation';
+import { createStatelessComponent } from '../../../Foundation';
 import { Stack } from '../Stack';
 import { StackItem } from '../StackItem/StackItem';
 import { IStackItemProps } from '../StackItem/StackItem.types';
 import { getVerticalAlignment, getHorizontalAlignment } from '../StackUtils';
-import { IVerticalStackProps, IVerticalStackStyles } from './VerticalStack.types';
+import { IVerticalStackComponent, IVerticalStackProps, IVerticalStackStyles } from './VerticalStack.types';
 import { styles } from './VerticalStack.styles';
 
-const view = (props: IViewComponentProps<IVerticalStackProps, IVerticalStackStyles>) => {
-  const { verticalAlign, horizontalAlign, classNames, children } = props;
+const view: IVerticalStackComponent['view'] = props => {
+  const { verticalAlign, horizontalAlign, gap, classNames, children, styles: customStyles, ...rest } = props;
 
   return (
     <Stack
-      {...props}
+      {...rest}
       verticalAlignment={getVerticalAlignment(verticalAlign)}
       horizontalAlignment={getHorizontalAlignment(horizontalAlign)}
       className={classNames.root}
+      verticalGap={gap}
     >
       {children}
     </Stack>
   );
 };
 
-const StackStatics = {
+const VerticalStackStatics = {
   Item: StackItem,
   defaultProps: {}
 };
-type IVerticalStackStatics = typeof StackStatics;
+type IVerticalStackStatics = typeof VerticalStackStatics;
 
 export const VerticalStack: React.StatelessComponent<IVerticalStackProps> & {
   Item: React.StatelessComponent<IStackItemProps>;
@@ -34,7 +35,7 @@ export const VerticalStack: React.StatelessComponent<IVerticalStackProps> & {
   displayName: 'VerticalStack',
   styles,
   view,
-  statics: StackStatics
+  statics: VerticalStackStatics
 });
 
 export default VerticalStack;