+++ packages/office-ui-fabric-react/src/components/Panel/Panel.base.tsx
@@ -1,24 +1,23 @@
 import * as React from 'react';
-
-import {
-  BaseComponent,
-  classNamesFunction,
-  divProperties,
-  getId,
-  getNativeProps,
-  getRTL,
-  createRef,
-  elementContains,
-  allowScrollOnElement,
-  isIOS
-} from '../../Utilities';
-import { IProcessedStyleSet, getTheme, IconFontSizes } from '../../Styling';
-import { FocusTrapZone } from '../FocusTrapZone/index';
-import { IPanel, IPanelProps, PanelType, IPanelStyleProps, IPanelStyles } from './Panel.types';
+import { IconButton } from '../../Button';
 import { Layer } from '../../Layer';
 import { Overlay } from '../../Overlay';
 import { Popup } from '../../Popup';
-import { IconButton } from '../../Button';
+import { getTheme, IconFontSizes, IProcessedStyleSet } from '../../Styling';
+import {
+  allowScrollOnElement,
+  BaseComponent,
+  classNamesFunction,
+  createRef,
+  divProperties,
+  elementContains,
+  getId,
+  getNativeProps,
+  getRTL,
+  isIOS
+} from '../../Utilities';
+import { FocusTrapZone } from '../FocusTrapZone/index';
+import { IPanel, IPanelProps, IPanelStyleProps, IPanelStyles, PanelType } from './Panel.types';
 
 const getClassNames = classNamesFunction<IPanelStyleProps, IPanelStyles>();
 
@@ -150,13 +149,7 @@ export class PanelBase extends BaseComponent<IPanelProps, IPanelState> implement
 
     let overlay;
     if (isBlocking && isOpen) {
-      overlay = (
-        <Overlay
-          className={_classNames.overlay}
-          isDarkThemed={false}
-          onClick={isLightDismiss ? onLightDismissClick : undefined}
-        />
-      );
+      overlay = <Overlay className={_classNames.overlay} isDarkThemed={false} onClick={isLightDismiss ? onLightDismissClick : undefined} />;
     }
 
     const header = onRenderHeader(this.props, this._onRenderHeader, headerTextId);
@@ -179,9 +172,7 @@ export class PanelBase extends BaseComponent<IPanelProps, IPanelState> implement
               className={_classNames.main}
               style={customWidthStyles}
               elementToFocusOnDismiss={elementToFocusOnDismiss}
-              isClickableOutsideFocusTrap={
-                focusTrapZoneProps && !focusTrapZoneProps.isClickableOutsideFocusTrap ? false : true
-              }
+              isClickableOutsideFocusTrap={focusTrapZoneProps && !focusTrapZoneProps.isClickableOutsideFocusTrap ? false : true}
             >
               <div ref={this._allowScrollOnPanel} className={_classNames.scrollableContent}>
                 <div className={_classNames.commands} data-is-visible={true}>
@@ -214,7 +205,7 @@ export class PanelBase extends BaseComponent<IPanelProps, IPanelState> implement
     }
   }
 
-  public dismiss = (ev?: React.KeyboardEvent<HTMLElement>): void => {
+  public dismiss = (ev?: React.SyntheticEvent<HTMLElement>): void => {
     if (this.state.isOpen) {
       if (this.props.onDismiss) {
         this.props.onDismiss(ev);
@@ -352,8 +343,8 @@ export class PanelBase extends BaseComponent<IPanelProps, IPanelState> implement
     }
   }
 
-  private _onPanelClick = (): void => {
-    this.dismiss();
+  private _onPanelClick = (ev?: any): void => {
+    this.dismiss(ev);
   };
 
   private _onTransitionComplete = (): void => {