+++ packages/styling/src/interfaces/ITheme.ts
@@ -2,8 +2,17 @@ import { IPalette } from './IPalette';
 import { IFontStyles } from './IFontStyles';
 import { ISemanticColors } from './ISemanticColors';
 import { ITypography, IPartialTypography } from './ITypography';
+import { ISpacing } from './ISpacing';
 
-export interface ITheme {
+/**
+ * @internal
+ * Predefined scheme identifiers.
+ * Schemes are is still in an experimental phase.
+ * This interface's naming and values are not finalized and are subject to change.
+ */
+export type ISchemeNames = 'default' | 'neutral' | 'soft' | 'strong';
+
+export interface IScheme {
   palette: IPalette;
   fonts: IFontStyles;
   semanticColors: ISemanticColors;
@@ -21,17 +30,35 @@ export interface ITheme {
 
   /**
    * @internal
-   * The typography property is still in an experimental phase. The intent is the have it
+   * The typography property is still in an experimental phase. The intent is to have it
    * eventually replace IFontStyles in a future release, but it is still undergoing review.
    * Avoid using it until it is finalized.
    */
   typography: ITypography;
+
+  /**
+   * @internal
+   * The spacing property is still in an experimental phase. The intent is to have it
+   * be used for padding and margin sizes in a future release, but it is still undergoing review.
+   * Avoid using it until it is finalized.
+   */
+  spacing: ISpacing;
+}
+
+export interface ITheme extends IScheme {
+  /**
+   * @internal
+   * The schemes property is still in an experimental phase. The intent is to have it work
+   * in conjunction with new 'schemes' prop that any component making use of Foundation can use.
+   * Alternative themes that can be referred to by name.
+   */
+  schemes?: { [P in ISchemeNames]?: IScheme };
 }
 
 export type IPartialTheme = {
   [P in keyof Pick<
     ITheme,
-    'palette' | 'fonts' | 'semanticColors' | 'isInverted' | 'disableGlobalClassNames'
+    'palette' | 'fonts' | 'semanticColors' | 'isInverted' | 'disableGlobalClassNames' | 'spacing' | 'schemes'
   >]?: Partial<ITheme[P]>
 } &
   { [P in keyof Pick<ITheme, 'typography'>]?: IPartialTypography };