+++ packages/variants/CHANGELOG.md
@@ -1,6 +1,24 @@
 # Change Log - @uifabric/variants
 
-This log was last generated on Thu, 13 Sep 2018 17:38:04 GMT and should not be manually modified.
+This log was last generated on Mon, 01 Oct 2018 12:27:24 GMT and should not be manually modified.
+
+## 6.11.0
+Mon, 01 Oct 2018 12:27:24 GMT
+
+### Minor changes
+
+- semantic slot value updates per design direction
+
+## 6.10.0
+Fri, 21 Sep 2018 14:25:46 GMT
+
+### Minor changes
+
+- Variants: export everything
+
+### Patches
+
+- Adding a version stamp file
 
 ## 6.9.0
 Thu, 13 Sep 2018 17:38:04 GMT